Semi-working Prototype
This is a semi-working prototype with a computer that effectively
selects a number, prints out certain things, and evaluates correctly if
the player wins or loses or should try again.

However, the round and game system could use some work; right now it
doesn't really work to go beyond round 1.