Stat printing
This commit has stats printed at the end of the game. Stats are printed
properly for round 1, and code has been refactored to make methods
more simple and easy to understand as well as make calculating stats
easy and nice. The calculations work and use methods from the different
encapsulated classes, but they can likely be refactored and simplified
more.

The calculation for shortest guess record needs work, as it needs to
happen after each round. This involves finding out where best to put
this calculation in the game, and likely moving it from game.java to
round.java. That will be next.