Stats properly printing
It needs more testing to make sure it works, but stats that exist are
properly printing when the game ends. The calculations seem to generally
be alright, so that's good.

Next step will be refactoring, adding comments, making sure that
everything is in the correct class for what it all intends to do;
aka, ENCAPSULATION!!!