Printing Guess List on Round Win + small refactoring
This commit involves printing the player's list of guesses when the
round is over, and it does it correctly with spacing and all that, so
it works :).

Otherwise, there was minor refactoring in terms of adding more comments,
but this will have to be continued into the next commit as well as
really adding more methods to simplify some of the semi-raw values in
the other methods. Then, it can be released and future updates made!