Refactoring tweaks + README
This commit is the final bit of minor refactoring for this version and
update. Of course, more refactoring can likely be done, but that will be
for future versions once I learn more about refactoring and making clean
+ efficient code and I revisit this project to add more bells and
whistles and a more engaging gameplay loop. For now, it is what it is.

This commit involved adding comments to functions and creating methods
so as not to use raw values in some conditionals. There are more that
can probably be caught, and I'm not sure if this is the best approach,
but it does make it look cleaner.

From here on out this is a working product! Anything else will be bug
fixing, refactoring, and adding new features.

Till then!