Stat printing (#1)
* Stat printing

This commit has stats printed at the end of the game. Stats are printed
properly for round 1, and code has been refactored to make methods
more simple and easy to understand as well as make calculating stats
easy and nice. The calculations work and use methods from the different
encapsulated classes, but they can likely be refactored and simplified
more.

The calculation for shortest guess record needs work, as it needs to
happen after each round. This involves finding out where best to put
this calculation in the game, and likely moving it from game.java to
round.java. That will be next.

* Stats properly printing

It needs more testing to make sure it works, but stats that exist are
properly printing when the game ends. The calculations seem to generally
be alright, so that's good.

Next step will be refactoring, adding comments, making sure that
everything is in the correct class for what it all intends to do;
aka, ENCAPSULATION!!!

* Printing Guess List on Round Win + small refactoring

This commit involves printing the player's list of guesses when the
round is over, and it does it correctly with spacing and all that, so
it works :).

Otherwise, there was minor refactoring in terms of adding more comments,
but this will have to be continued into the next commit as well as
really adding more methods to simplify some of the semi-raw values in
the other methods. Then, it can be released and future updates made!

* Refactoring tweaks + README

This commit is the final bit of minor refactoring for this version and
update. Of course, more refactoring can likely be done, but that will be
for future versions once I learn more about refactoring and making clean
+ efficient code and I revisit this project to add more bells and
whistles and a more engaging gameplay loop. For now, it is what it is.

This commit involved adding comments to functions and creating methods
so as not to use raw values in some conditionals. There are more that
can probably be caught, and I'm not sure if this is the best approach,
but it does make it look cleaner.

From here on out this is a working product! Anything else will be bug
fixing, refactoring, and adding new features.

Till then!