GUI
This first GUI commit adds the game window wrapper class for JFrame, including a set width and height, default close operation, and not resizable. There are obviously some formatting and other cool things that
can be done, but for now it's going to be this. This commit is the frame wrapper constructor and that's about it.