INPUT ON TO SCREEN
There is now input on to the central screen from the text field, allowing the player to input text that then gets added to the text log. There are still formatting things that can be done, but the text wraps and
the input works! So I think for now I'll move on to genuine functionality and refactoring.