PRINTING MESSAGES
This commit adds the initial parts of the architecture for printing messages. It can certainly be refactored, perhaps using hash maps or some other dictionary that would make things a hell of a lot easier,
but for now this (might) work. I'll have to test it, but I can also condense this list to a much smaller one which will be nice, and it's relatively easy to add to should I need to.

One thing I can do here for future is add a CUSTOM type that just takes in the vars[0] argument, and then have local variables that define that wherever they are used. That would help shorten the list of
cases when no vars are being called upon (and remember--these vars are in String form, so any ints must be converted, etc.).