PRINT MESSAGE WORKS ON INPUT
This commit is really cool; the print message object now works on input!!! The game now logs the custom Print Message object and converts it to a string that appears as a normal string does. This was
done through a toString of the Print Message, which returns it as a string value--it takes all values of the array and prints them, which for the input is just the custom type (although I may make a
separate input type, there may be no need since custom works for it and custom strings. If no custom strings come about, I'll just rename it to the input type).

It also adds a line break afterward--something good to know if things break.

The other things printed by an array to string (using the Arrays.toString() method) are replaced with blank spaces, so no commas or brackets.