DIFFICULTY DIALOG BOX
This commit adds a GUI difficulty dialog box, which will take in a button press from the player on what difficulty they want and the game controller will apply it to the game state. That application functionality
has not yet been added, but the next commit will do that. The way I'll test it is that the guessed number should change based on what the player selected.