DIFFICULTY DIALOG FUNCTIONALITY ADDED
This commit finishes the difficulty dialog box, adding its functionality. The controller now takes in the difficulty choice from the dialog box and applies it to the game state, then takes that value from the
game state and displays the proper number of guesses on the game log in the printing difficulty situation. This framework of GUI gives something to controller that gives it to game state and then controller uses
the state of game state to print something in GUI seems pretty good MVC! There is that extra step, but it keeps generally clean architecting.