RUNNING THE ROUND
This is the first commit of this branch which will be all about running the round with the GUI integrated. This won't run the whole game yet, but just the round; the whole game will come later. Right now, the
GUI is able to read from the new round that is created by the controller and display the correct information regarding that round, including the maximum number of guesses for that round and the round number.

The next steps will be dealing with actually getting the Player and Computer (particularly the computer with all its logic) into the system as well.

In the future, what I'd like to do is have a menu bar that shows a dropdown with the stats of the player, but for now we'll print it at the end of each round.