MOVED THE STRINGS
Created a separate class for strings, but might move them into game view and just have methods there that deal with printing declarations, I'm not too sure. Either way, this was a minor little change, and next is
figuring out how to stop the GUI from going while it waits for player input; likely by adding that to the action listener on the inputter.