ALLOWING ONLY NUMBERS
Using a try catch block, the code tries to parse the inputted value in the field into an Int. If it fails, it catches and labels it an invalid input, clearing the text field for more input and printing that the
input was invalid without attributing that value to any variable. Thus, the player is limited to only inputting integers.