PLAYER GUESS INFRASTRUCTURE
This commit adds infrastructure for the controller to access the player guess as well as the GUI to create a local variable to store the current input, which the controller applies to the player and uses to run
its evaluation. The only thing that must be done next is making it so that the action listener to the input field is created in the game controller rather than the game view, since it needs to access the player
and everything to run the next phase of things after the guess has been inputted.

Things generally do need to be cleaned up, with printing itself being put into game view and the controller merely calling upon those methods. But that can be fixed later.