MOVED THE ACTION LISTENER
This commit moves the action listener for the inputter from game view, where it was unable to access any game state data, to the controller, where it can now control the flow of the GUI based on game state data.

This commit also moves the printing method into a centralized logInput method in game view for it to log either a successful or unsuccessful input. Other methods will be refactored to be the same, combining
where possible, and also hopefully being able to create a centralized custom print method (the print system kinda isn't working the way I thought but whatever).