GUESS COMPARISON IN GUI
This commit adds the guess comparison to the Game Controller and GUI, which in turn means that the game in the GUI now works all the way up to letting the player guess the number and displaying if they are
lower or higher! There is no ending of the round, but the round is able to run in the GUI!

Next I will work on ending the round as a win or a loss, then after that will be the game stuff in between rounds. I'm thinking a dialog box that asks about continuing and prints stats. Then, it will be all
about refactoring and making things cleaner, more encapsulated and separated, and BETTER!!!!