ENDING THE ROUND
This branch will be focused on how to end the round. Perhaps booleans in the game controller (or booleans read from the game state would be better) is the best way for me to determine when the round is over and
what should be printed. So I'll do that, but that will be tomorrow. I wanted my foot in the door with a commit so I remembered where I was, and so I had a nice streak on GitHub activity :)