ENDING ROUND + PLAYER GUESS
The guess framework has been refactored to be much better. Now, the game takes in the input from the controller, and then runs the calculation of game state on it, and then gives it back to the controller for
GUI printing. This is done through returning an enum based on what happens from the player's guess; they win, or lose, or guess again with a hint depending on the situation.

The round also does effectively end when the player's guesses exceed the maximum, as the game controller prints based on whether that is calculated to be the case by the game state.

Now, the actual removing of input when the loss or win occurs must happen, and proper exiting/continuing functionality! So this branch continues.