ROUND OVER DIALOG BOX CREATED
This commit creates a class for the round over dialog box which will show up when the player either loses or wins the round. It will display the game stats. If the player wins, it will ask them if they would like
to continue. If the player loses, it will tell them that the game is now exiting, with a button for the player to say "okay" when they are ready.