GAME DIALOG SUPER CLASS + ABSTRACTED CODE
This commit, in preparation for the dialog box for ending the round, abstracts out code from dialog boxes to make them more centralized and make it easier to make individual ones. For instance, it creates the
buttons and adds them all in the superclass, so all that needs to be done is creating the JPanel in each subclass (or, in the future, making all subclasses have a buttonPanel if they all will have one) and
giving the buttons their proper action listeners.

Now, I can move on to actually making the round end dialog box!