ROUND OVER DIALOG COMPLETED + ABSTRACTION OF LAYOUT AND BUTTON PANEL
This commit abstracts the logic to create and add the button panel, as well as to format the dialog box properly, to the game dialog instead of the individual subclasses.

This commit also completes the RoundOverDialog class, which will bring up a dialog box when the round is completed and print the array list of guesses. It will also have to print the strings attached to that
message, so perhaps I will have to do some finagling with that.

However, it is roughly ready. I'll add all the strings and then it will be ready for implementation.