ROUND OVER DIALOG TEXT DISPLAY
This commit displays all the text required for the round over dialog. The formatting is messed up which can be fixed later, but it is functional. The buttons are also hard coded here, which I don't really like,
but that might just be the best option--besides, the only place yes or no comes up is here anyway, so it's fine for now and can be fixed later.

Next, the action listeners need to be added to the buttons! I'd like for it to change some boolean in the round over dialog that the game controller reads; a continueGame boolean. That's the best option for sure.
Then, GameController does its stuff based on that input.