ROUND OVER SORTA KINDA WORKING?
The round over logic is now sort of kind of working! It's incredibly bare bones, DESPERATELY needs to be refactored, and needs to be formatted, but now this commit adds the round dialog box in the game controller
with a boolean that is modified when one of the buttons are pressed, applies that to the game state, and now does logic based on that game state value. The system exits when the player clicks no, and sets a log
message when the player clicks yes. Now I just need to refactor and make better, and I'm good to go!

However this branch ends here. Now I'm going to refactor the Strings code because that desperately needs it. It's way too unwieldy. So that'll be another branch.