ENUM MAP OF GAME STRINGS
This commit adds an enum map of game strings to centralize all the strings and pair them with enums rather than using indexed arrays which is just... so much better. So much better. Now the strings can be
referenced by their enum rather than by an index in an array. Next, we need to implement this new system. I might also refactor where the print data is and how it works, maybe just centralizing it all in the game
GUI where the text log is, although it would be nice to store a reference to that game log in some way.

Anyway, on to the actual implementation steps for the next commit.