STRINGS REFACTORED -- NO MORE PRINT MESSAGE TYPES
The strings have been condensed, so there is only an enum for game strings now which is applied as keys to an enum map of all the strings in the game.

I'm not sure which was better; having the print message type with all the string data in the enums themselves, or having a game strings hashmap with the types just as non-parametered types and just having the
hashmap hold all the strings. But I guess we can see how it works. It seems to reduce the code bloat a bit in terms of what's in Game Controller and otherwise, which is good and what I wanted, so that will
probably be best.

Now it's gonna be all about refactoring the actual game code!