REFACTORING THE DIALOG BOXES
This commit refactors the dialog boxes to make the code cleaner and hopefully a little better. They still need to be greatly refactored but this at least helps.

One of the main issues I need to figure out now is the formatting and how to make it so that new lines are created between text in the JLabel, which honestly won't be too bad; some HTML apparently between the
lines can work.