REFACTORING GAME CONTROLLER AND VIEW
This commit refactors the game controller class and game view class, which were both much overdue for a refactor. All the actual printing was moved to the game view, where the controller just calls on those
methods and does the logic, passing in any parameters needed. This makes both classes much cleaner and creates a much better separation of responsibilities between them.

This commit also introduced and fixed bugs, like the parsing of int and the try catch block involving displaying input, and made the action listener much easier to read and understand. There are still definitely
some refactors that can be made but this is overall much cleaner.