BUG FIXING AND REFACTORING (#10)
* BUG FIXING AND REFACTORING

This commit cleans up the game code by removing unused things, print statements, etc., and makes room for better game code that can involve system logging, error handling, etc.

* REFACTORING THE DIALOG BOXES

This commit refactors the dialog boxes to make the code cleaner and hopefully a little better. They still need to be greatly refactored but this at least helps.

One of the main issues I need to figure out now is the formatting and how to make it so that new lines are created between text in the JLabel, which honestly won't be too bad; some HTML apparently between the
lines can work.

* REFACTORING GAME CONTROLLER AND VIEW

This commit refactors the game controller class and game view class, which were both much overdue for a refactor. All the actual printing was moved to the game view, where the controller just calls on those
methods and does the logic, passing in any parameters needed. This makes both classes much cleaner and creates a much better separation of responsibilities between them.

This commit also introduced and fixed bugs, like the parsing of int and the try catch block involving displaying input, and made the action listener much easier to read and understand. There are still definitely
some refactors that can be made but this is overall much cleaner.

* REMOVING TO-DOs

This commit just removes some to-dos and flags from these classes.

* FIXING MINOR BUG + FUTURE PLANS

This commit fixes the minor bug in game controller where invalid inputs never let the text field accept inputs ever again. This commit also adds future plans to the read me, including a schedule for the coming
days that will be updated once all the goals have been completed.