GAME OVER DIALOG
This commit creates the game over dialog box class and does a little bit of work with it like creating the okay button. The formatting is all messed up as per usual, but it does actually work. The okay button
on it just exits the game, so now I actually have to implement this dialog box when the player selects no on the round over dialog.

Dialog boxes into dialog boxes....