GAME OVER DIALOG
This commit finishes up and adds proper functionality to the game over dialog. There are bugs that need to be fixed, but it is there. I think I'm going to take a week off from this project; I'm feeling burnt out
from it, as awesome as it was learning so much about Swing, as it's not really one that feels on the cutting edge of what I'm learning. At least, not yet. When I start learning more things, I'm going to add them
here and use this as an experiment of sorts. But for now, the game over dialog is added and good, so I'll commit this branch to main and create a new release.